Fixing deletion issue (container, client)

+++ src/main/java/com/crudCommerce/service/ClienteService.java
@@ -9,12 +9,16 @@ import org.springframework.stereotype.Service;
 import com.crudCommerce.exception.RegraNegocioException;
 import com.crudCommerce.model.Cliente;
 import com.crudCommerce.repository.ClienteRepository;
+import com.crudCommerce.repository.ConteinerRepository;
 
 @Service
 public class ClienteService {
 	
 	@Autowired
 	private ClienteRepository repository;
+	
+	@Autowired
+	private ConteinerRepository conteinerRepostory;
 	
 	public List<Cliente> listAllClientes() {
 		return this.repository.findAll();
@@ -34,6 +38,12 @@ public class ClienteService {
 	}
 	
 	public void deleteCliente(Integer id) {
+		Optional<Cliente> cliente = findClienteById(id);
+		
+		if(this.conteinerRepostory.existsByCliente(cliente.get())) {
+			throw new RegraNegocioException("Não é possivél deletar este cliente pois existe um contêiner cadastrado com este cliente");
+		}
+		
 		this.repository.deleteById(id);
 	}
 }
